add a stacktemp helper function to the sandbox
The intent of this function is to get the same effect one would get with
`stackarr`, but without the need for an existing (non-sandboxed)
temporary and the somewhat unintuitive size argument.  Using this in
preference to `mallocInSandbox` also ensures that you don't forget to free
the malloc'd buffer--although you do have to exercise good judgement
about using this function.